refactor: Rename project from `barberflow-api` to `inkflow-api`
