--- src/main/java/br/com/barberflow/api/core/user/domain/UserRole.java
@@ -1,7 +1,0 @@
-package br.com.barberflow.api.core.user.domain;
-
-public enum UserRole {
-    ADMIN,
-    BARBER,
-    CLIENT;
-}
