+++ src/main/java/br/com/inkflow/api/HealthCheckController.java
@@ -1,4 +1,4 @@
-package br.com.barberflow.api;
+package br.com.inkflow.api;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
