+++ src/main/java/br/com/inkflow/api/SecurityConfig.java
@@ -1,4 +1,4 @@
-package br.com.barberflow.api;
+package br.com.inkflow.api;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.Customizer;
