+++ src/main/java/br/com/inkflow/api/core/user/domain/UserRole.java
@@ -1,0 +1,7 @@
+package br.com.inkflow.api.core.user.domain;
+
+public enum UserRole {
+    ADMIN,
+    TATTOOIST,
+    CLIENT;
+}
