+++ src/test/java/br/com/inkflow/api/InkflowApiApplicationTests.java
@@ -1,10 +1,10 @@
-package br.com.barberflow.api;
+package br.com.inkflow.api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BarberflowApiApplicationTests {
+class InkflowApiApplicationTests {
 
 	@Test
 	void contextLoads() {
